Copied contents back from BlueprintFunctions-fix.js to BlueprintFunctions.js
